Move jwt-token and token(creator) into cmds and update Make and Docker
